New LayoutManager implementation.
Store the hosts grid representation and stop trying to compute host position dynamically.
This greatly simply and fix issues with the selection API (select next/previous) which was assuming that all rows where full but the last one. Unfortunately, that is not longer true with the "force rows count" API.