Slow down the death of host.
If an host die before the controller did resolved it pid and tty, the later cannot define while HostWindow should be closed.
To avoid that, introduce a handshake byte sent by the controller to each host on connection. If SSH dies before the handshake, wait
until the handshake is done to close the connection and exit.